+++ rcon/auto_settings.py
@@ -8,7 +8,7 @@ import pytz
 from rcon.audit import ingame_mods, online_mods
 from rcon.rcon_ import Rcon
 from rcon.settings import SERVER_INFO
-from rcon.user_config.user_config import AutoSettingsConfig
+from rcon.user_config.auto_settings import AutoSettingsConfig
 from rcon.vote_map import VoteMap
 
 logger = logging.getLogger(__name__)
