+++ rcon/user_config/auto_settings.py
